[mod/#4] 패키지 구조 변경 및 빈 화면 구현
