+++ app/src/main/java/com/example/letssopt/presentation/home/component/HomeUpComingSection.kt
@@ -1,4 +1,4 @@
-package com.example.letssopt.presentation.main.component
+package com.example.letssopt.presentation.home.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
@@ -13,16 +13,13 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.example.letssopt.R
@@ -30,8 +27,9 @@ import com.example.letssopt.designsystem.theme.LETSSOPTTheme
 import com.example.letssopt.designsystem.theme.WatchaTheme
 import com.example.letssopt.model.MainHomeItem
 
+
 @Composable
-fun AlgorithmSection(
+fun UpComingSection(
     contents: List<MainHomeItem.Contents>,
     modifier: Modifier = Modifier,
 ) {
@@ -45,19 +43,11 @@ fun AlgorithmSection(
             horizontalArrangement = Arrangement.SpaceBetween,
             verticalAlignment = Alignment.CenterVertically,
         ) {
-            Column {
-                Icon(
-                    ImageVector.vectorResource(R.drawable.ic_watcha_algorithm),
-                    null,
-                    tint = WatchaTheme.colors.textPrimary
-                )
-                Spacer(modifier = Modifier.height(4.dp))
-                Text(
-                    text = "예능부터 드라마까지",
-                    style = WatchaTheme.typography.cap.captionR20,
-                    color = WatchaTheme.colors.textSecondary,
-                )
-            }
+            Text(
+                text = "공개 예정 콘텐츠",
+                style = WatchaTheme.typography.headline.head2B20,
+                color = WatchaTheme.colors.textPrimary,
+            )
             Text(
                 text = "더보기",
                 style = WatchaTheme.typography.body.bodyR12,
@@ -65,7 +55,7 @@ fun AlgorithmSection(
             )
         }
 
-        Spacer(modifier = Modifier.height(6.dp))
+        Spacer(modifier = Modifier.height(8.dp))
 
         LazyRow(horizontalArrangement = Arrangement.spacedBy(16.dp))
         {
@@ -86,14 +76,13 @@ fun AlgorithmSection(
 
 @Preview
 @Composable
-private fun AlgorithmSectionPreview() {
+private fun UpcomingSectionPreview() {
     LETSSOPTTheme {
-        AlgorithmSection(
+        UpComingSection(
             contents =
                 listOf(
                     MainHomeItem.Contents("이 사랑 통역 되나요?", R.drawable.img_poster_love_translate),
                     MainHomeItem.Contents("기묘한 이야기", R.drawable.img_poster_stranger_things),
-                    MainHomeItem.Contents("프로젝트 헤일메리", R.drawable.img_poster_hail_mary),
                 )
         )
     }
