+++ app/src/main/java/com/example/letssopt/presentation/search/SearchScreen.kt
@@ -1,0 +1,37 @@
+package com.example.letssopt.presentation.search
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun SearchRoute(
+    paddingValues: PaddingValues
+) {
+    SearchScreen(
+        paddingValues = paddingValues
+    )
+}
+
+@Composable
+fun SearchScreen(
+    paddingValues: PaddingValues
+) {
+    Column (
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(paddingValues),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Text(
+            text = "Search"
+        )
+    }
+}
