+++ app/src/main/java/com/example/letssopt/presentation/webtoon/WebtoonScreen.kt
@@ -1,0 +1,38 @@
+package com.example.letssopt.presentation.webtoon
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import com.example.letssopt.presentation.search.SearchScreen
+
+@Composable
+fun WebtoonRoute(
+    paddingValues: PaddingValues
+) {
+    SearchScreen(
+        paddingValues = paddingValues
+    )
+}
+
+@Composable
+fun WebtoonScreen(
+    paddingValues: PaddingValues
+) {
+    Column (
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(paddingValues),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Text(
+            text = "Webtoon"
+        )
+    }
+}
